Added test of moving players about within teams after teams had already been allocated.
This was added as part of existing test rejoin-restored-game-middle-of-round, by adding the possibility to put custom actions into a test as part of the game-defs.

Also simplified some of the functions in common-test-functions by passing the args as an object instead of individually.

Did a fix in script.js: hideAllContextMenus() wasn't always being called when it was added as part of a second click event listener. Had to instead add it to the end of all the existing click event listeners.

Removed some traces that were getting in the way.