Added code both server-side and client-side to prevent a player starting a turn more than once (previously this could happen, resulting in multiple countdown timers being created).
Includes more generic error-handling mechanism, using IllegalServerRequestException and the ILLEGAL_REQUEST error code. On the client side, the button is simply hidden as soon as it's clicked.

Problem: the way this actually occurred was that the websocket had closed, so the StartTurn request was processed, but the client didn't receive the result. Now, if this occurs, the client still won't see the result, although the StartTurn button will be hidden. The user just needs to refresh, but this won't be obvious.