Added function addServerRequestClickListener to have uniform handling of click event listeners on buttons which send a request to the server (e.g. when to disable/enable, to ensure I never forget to call restoreWebSocketIfNecessary(), and input validation).
For now, not being used on startNextRoundButton and startTurnButton, because the new function doesn't have a natural way to include the call to clearTestTrigger(), and without that the tests don't pass.

Added input validation on game params. Also added some doc comments in api-calls.js.

'restore' cookie is no longer used when restoring the websocket. When a new websocket is opened, we just always restore. Means the behaviour on opening for the first time is more consistent with any re-opening - necessary since the nameSubmitButton no longer has its own custom click event listener, so the websocket is opened from there with restoreIfNecessary, and not with tryToOpen as before.