Previous attempted fix of the delay, and the margin of 10s being too low, wasn't enough. Now trying reading the seconds remaining from the page.
Had to change the code which takes the turns to a recursive method rather than a loop, so that each successive iteration can be called from a then(). This allows us to read the seconds remaining from the page, before deciding how long to wait before taking the next move.