Made the check of the scoresDiv at the end of each round much more efficient. Before, it was calling cy.get('[id="scoresDiv"]') once for each celebName - cy.get() is a very slow operation on the DOM!
Now, it calls it just once, and parses the text of the scoresDiv. As a bonus, this checks extra things, such as that extra text isn't included in the div.

In doing so, also made it work the same way with the check at the end of each turn. It turned out that the cy.get()-based code for this check wasn't being done at all for this, because of the usual annoying variable scope/execution order issues in cypress. So, although the check has now been made quicker, the overall test probably runs slower. That said, it is now checking more things, and doing it more efficiently.