Instead of re-loading the page when being told a request was illegal, just show the notification and restore the websocket if necessary.
Unless you're using the console to force an illegal request, the usual way to make one is if the websocket has closed, and a button which should have been hidden is still visible. The only fix that's needed is to restore the websocket, which will result in the latest game state being provided - refresh would accomplish the same thing, but is overkill.

Made api-calls.js a module, to be able to call showNotification from there. Made 'exporting' stuff via the window object a bit more systematic, by creating window.hackExport.