Fix: pings weren't handled correctly (or at all). We reached the line which logs the message "Unexpected byte" - at this point we also stop the whole WebsocketHandler.
Have kept this fail-fast behaviour for now, even though we don't really need to stop the handler just for that. Up until now it wasn't a problem, and it helped me notice something new was happening. Started getting pings in the Cypress tests, although only when testing on the live server, and only when using Firefox. Don't know why it didn't happen before.

The pings came in 10-byte websocket frames, encoded in the usual way. The bytes were:
0: PING_BYTE
1: length indicator, indicating a length of 4
2-5: key
6-9: the ASCII string "PING", encoded with the key

After checking RFC 6455, including a message with your ping is optional. If you do, the pong must include the same message. This is now implemented, and we log the pings and pongs we receive. Don't necessarily need to keep this logging behaviour for long, but it'll be interesting to see if they keep happening.