Only create a Session after somebody has entered a username. Will make the logs easier to read and follow, since we don't create a Session and a Player for every bot who downloads the page.
Because of this, need to open Websocket slightly later (websocket must be associated to a session). Function restoreWebsocketIfNecessary(), which is called after every button click (including the one to send the username), now checks if there is a session cookie first. It can still happen that we try to open it too early (e.g. if we restart the server, and the browser remembers a session ID from a previous instance). If so no big deal, it will just be opened correctly on the next button click. But in normal operation, we will arrive without a session, and not try to open the websocket before getting one.

Username button now calls tryToOpenSocket after it has received a response to its fetch request (function added as arg to addServerRequestClickListener).