Started work on getting new server code to handle websockets as well.
For now, the plan is to still have them handled by the same HTTPChannelHandler, since the initiation comes from an HTTP request, and then we keep that TCP connection going. That code isn't complete yet, but there is code in OpenWebsocketRequestHandler to send the appropriate handshake response, and to mark the corresponding SocketChannel as a separate websocket channel in the HTTPChannelHandler.

For the handling of incoming messages, created class WebsocketMessage, analagous to HTTPExchange, and added a test class for it. Code in common between WebsocketMessage and the old WebsocketHandler has been moved to new class WebsocketUtil.