Modified the Allocate Teams request to include a numTeams argument based on the value selected in the drop-down box.
This confirms the drop-down box should always contain a value, we shouldn't only start putting a value when there are 4 or more players (for testing we sometimes use 1 player).

Added new field allPlayers to JSON representation of Game, including players both in and out of teams, to make it easier to calculate the possible numbers of teams when re-allocating.

Also made the drop-down box keep whichever value was previously selected, after updating (if possible).