Got cypress tests working with  more than 2 teams.
numTeams is read from environment (test script was already setting it), and host selects the appropriate value from the drop-down. Places in the testing code that had been hard-coded to expect 2 teams were modified. Not much to change, as I'd had in mind eventually that this would eventually be supported. Main 2 things were:
- calculating the turnIndex in the pre-set turns
- reading the scores div